+++ src/main/webapp/wise5/directives/component/component.es6
@@ -79,7 +79,12 @@ class ComponentController {
             $scope.nodeAuthoringController = $scope.$parent.nodeAuthoringController;
         }
 
-        if ($scope.mode === 'authoring' && (componentContent.type == 'HTML' || componentContent.type == 'ConceptMap' || componentContent.type == 'Animation' || componentContent.type == 'Graph' || componentContent.type == 'Label' || componentContent.type == 'Match' || componentContent.type == 'OpenResponse' || componentContent.type == 'Table' || componentContent.type == 'OutsideURL')) {
+        if ($scope.mode === 'authoring' &&
+            (componentContent.type == 'HTML' || componentContent.type == 'ConceptMap' ||
+              componentContent.type == 'Animation' || componentContent.type == 'Graph' ||
+              componentContent.type == 'Label' || componentContent.type == 'Match' ||
+              componentContent.type == 'OpenResponse' || componentContent.type == 'Table' ||
+              componentContent.type == 'OutsideURL' || componentContent.type == 'MultipleChoice')) {
             $scope.componentTemplatePath = this.NodeService.getComponentAuthoringTemplatePath(componentContent.type);
         } else {
             $scope.componentTemplatePath = this.NodeService.getComponentTemplatePath(componentContent.type);
